+++ R/make_attribute_topology.R
@@ -1,8 +1,11 @@
 #' @title Make Attribute Topology
 #' @description given a set of lines with starting and ending nodes that
 #' form a geometric network, construct an attribute topology.
-#' @inheritParams add_levelpaths
+#' @param x data.frame network compatible with \link{hydroloom_names}.
 #' @details
+#'
+#' Required attributes: `id` and sf linestring geometry
+#'
 #' If a `future` plan is set up, node distance calculations will be
 #' applied using future workers.
 #'