+++ R/utils.R
@@ -217,13 +217,13 @@ fix_flowdir <- function(id, network = NULL, fn_list = NULL) {
 }
 
 #' @title Rescale Aggregate id Measure to id Measure
-#' @description Given a aggregate id (e.g. reachcode) measure and the from and
-#' to measure for a id (e.g. comid flowline), returns the measure along the
-#' id line. This is a utility specific to the NHDPlus data model where
-#' many comid flowlines make up a single reachcode / reach. "Measures"
-#' are typically referenced to reaches. Flowlines have a stated from-measure /
-#' to-measure. In some cases it is useful to rescale the measure such that it
-#' is relative only to the flowline.
+#' @description Given an aggregate_id (e.g. reachcode in NHDPlus) measure and
+#' the from and to measure for an id (e.g. COMID flowline in NHDPlus), returns
+#' the measure along the id flowline. This is useful where many flowlines make
+#' up a single aggregate feature. "Measures" are typically referenced to the
+#' aggregate feature. Flowlines have a stated from-measure / to-measure. In
+#' some cases it is useful to rescale the measure such that it is relative only
+#' to the flowline.
 #'
 #' from is downstream -- 0 is the outlet
 #' to is upstream -- 100 is the inlet
@@ -304,15 +304,17 @@ force_linestring <- function(x) {
   x
 }
 
-add_toids_internal <- function(x, var = NULL) {
+add_toids_internal <- function(x, var = NULL, keep = FALSE) {
 
   if (all(c(id, fromnode, tonode, divergence) %in% names(x)) &&
     !toid %in% names(x)) {
 
     x |>
-      st_drop_geometry() |>
-      select(any_of(c(id, fromnode, tonode, divergence, as.character(var)))) |>
-      add_toids(return_dendritic = FALSE)
+      st_drop_geometry()
+
+    if (!keep) x <- select(x, any_of(c(id, fromnode, tonode, divergence, as.character(var))))
+
+    add_toids(x, return_dendritic = FALSE)
 
   } else {